Moved `PlotControl.py` to new module `Plotting`
