+++ LabExT/View/Controls/Plotting/PlotControl.py
